streamer/consumer_to_db/requirements.txt
json5==0.9.5
kafka-python==2.0.2
psycopg2==2.8.6
PyYAML==5.4.1
Telethon==1.23.0
regex==2021.9.24
pytz==2020.1